# openapi.yaml
openapi: 3.0.0
info:
  title: Conference Discovery & Research API
  description: REST API for discovering academic conferences and enriching them with research paper data
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Author:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        h_index:
          type: integer
        semantic_scholar_id:
          type: string
        affiliation:
          type: string
        last_updated:
          type: string
          format: date-time
    Paper:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        conference_id:
          type: integer
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
...
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
paths:
...
  /authors:
    get:
      tags:
        - Authors
      summary: List all authors
      description: Get paginated list of authors, optionally filtered by name, sorted by h-index
      parameters:
        - name: name
          in: query
          description: Search author by name
          schema:
            type: string
          example: John Smith
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Results per page (1-100)
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Authors list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PaginatedResponse'
  /authors/{author_id}:
    get:
      tags:
        - Authors
      summary: Get author details
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      author:
                        $ref: '#/components/schemas/Author'
                      papers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Paper'
        '404':
          description: Author not found
  /authors/{author_id}/refresh:
    post:
      tags:
        - Authors
      summary: Manually refresh author h-index (Admin only)
      description: Fetch fresh h-index data from Semantic Scholar
      security:
        - bearerAuth: []
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author h-index refreshed
        '403':
          description: Admin access required
        '404':
          description: Author not found
  /authors/{author_id}/papers:
    get:
      tags:
        - Authors
      summary: Get papers by author
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Results per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Papers retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      author:
                        $ref: '#/components/schemas/Author'
                      papers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Paper'
        '404':
          description: Author not found
...